+++ src/main/java/edu/harvard/iq/dataverse/api/AbstractApiBean.java
@@ -77,7 +77,7 @@ public abstract class AbstractApiBean {
 
     private static final Logger logger = Logger.getLogger(AbstractApiBean.class.getName());
     private static final String DATAVERSE_KEY_HEADER_NAME = "X-Dataverse-key";
-    private static final String PERSISTENT_ID_KEY=":persistentId";
+    protected static final String PERSISTENT_ID_KEY=":persistentId";
     private static final String ALIAS_KEY=":alias";
     public static final String STATUS_WF_IN_PROGRESS = "WORKFLOW_IN_PROGRESS";
     public static final String DATAVERSE_WORKFLOW_INVOCATION_HEADER_NAME = "X-Dataverse-invocationID";
@@ -242,6 +242,9 @@ public abstract class AbstractApiBean {
     
     @EJB 
     GuestbookResponseServiceBean gbRespSvc;
+    
+    @EJB 
+    TemplateServiceBean templateSvc;
 
     @Inject
     FailedPIDResolutionLoggingServiceBean fprLogService;
@@ -370,8 +373,18 @@ public abstract class AbstractApiBean {
         }
         return dv;
     }
+    
+    protected Template findTemplateOrDie(Long templateId) throws WrappedResponse {
+        
+        Template template = templateSvc.find(templateId);
+        if (template == null) {
+            throw new WrappedResponse(
+                    error(Response.Status.NOT_FOUND, "Can't find template with identifier='" + templateId + "'"));
+        }
+        return template;
+    }
 
-    protected Template findTemplateOrDie(Long templateId, Dataverse dataverse) throws WrappedResponse {
+    protected Template findTemplateInDataverseOrParentsOrDie(Long templateId, Dataverse dataverse) throws WrappedResponse {
         
         List<Template> templates = new ArrayList<>();
         
@@ -488,7 +501,27 @@ public abstract class AbstractApiBean {
     }
 
     protected void validateInternalTimestampIsNotOutdated(DvObject dvObject, String sourceLastUpdateTime) throws WrappedResponse {
-        Date date = sourceLastUpdateTime != null ? DateUtil.parseDate(sourceLastUpdateTime, "yyyy-MM-dd'T'HH:mm:ss'Z'") : null;
+        // The timestamp string must always be in UTC, ISO 8601-formatted
+        // for example: 2026-04-22T14:30:00Z. This is explicitly specified in the
+        // API guide.
+        //
+        // In the intended workflow, the clients will be reusing the last update
+        // timestamps obtained from the output of other Dataverse APIs, such as
+        // /versions and /files, where they are always in that form, regardless
+        // of the actual time zone the server lives in.
+        //
+        // For consistency, we do not want to accept any other formats or timezones,
+        // and will reject anything that does not match "yyyy-MM-dd'T'HH:mm:ss'Z'".
+        // For that reason there is an explicit check added for .endsWith("Z").
+        // The "X" in the parsing format string will recognize literal 'Z' as "+0000",
+        // but it would also accept other ISO 8601 timezones, such as "+0400" for
+        // EDT, etc. In theory, we could accept all these other notations - in
+        // case the client decided to convert the UTC timestamp they received from
+        // Dataverse into that... but there is really no good reason to encourage
+        // that.
+        Date date = sourceLastUpdateTime != null && sourceLastUpdateTime.endsWith("Z")
+                ? DateUtil.parseDate(sourceLastUpdateTime, "yyyy-MM-dd'T'HH:mm:ssX")
+                : null;
         if (date == null) {
             throw new WrappedResponse(
                     badRequest(BundleUtil.getStringFromBundle("jsonparser.error.parsing.date", Collections.singletonList(sourceLastUpdateTime)))